requirements.txt
playwright==1.57.0
gradio==4.44.1
openai==1.61.0
beautifulsoup4==4.13.3
python-dotenv==1.0.1
